Changed poms to apply Spotless pluging during compilation
